requirements.txt
opencv
numpy
matplotlib
keras
tensorflow
pandas
pillow
skimage
seaborn
pathlib
sklearn